# docs/swagger.yaml
basePath: /api/
definitions:
...
  10_0_0_50_tuan_quang_tran_vms-v2_internal_models.ContentReport:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      media_id:
        type: string
      media_type:
        type: string
      reason:
        $ref: '#/definitions/10_0_0_50_tuan_quang_tran_vms-v2_internal_models.ContentReportReason'
      reporter_ip:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  10_0_0_50_tuan_quang_tran_vms-v2_internal_models.ContentReportReason:
    enum:
      - sexual_content
      - violence
      - hateful_or_abusive_content
      - harassment_or_bullying
      - harmful_or_dangerous_acts
      - misinformation
      - child_abuse
      - promotes_terrorism
      - spam_or_misleading
      - copyright
      - captions_issue
    type: string
    x-enum-varnames:
      - ReasonSexualContent
      - ReasonViolence
      - ReasonHatefulOrAbusive
      - ReasonHarassmentOrBullying
      - ReasonHarmfulOrDangerous
      - ReasonMisinformation
      - ReasonChildAbuse
      - ReasonPromotesTerrorism
      - ReasonSpamOrMisleading
      - ReasonCopyright
      - ReasonCaptionsIssue
...
  ResponseError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  ResponseSuccess:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_controllers.CreateReportContentRequest:
    properties:
      description:
        type: string
      media_id:
        type: string
      media_type:
        type: string
      reason:
        type: string
    type: object
  internal_controllers.GetReportContentListData:
    properties:
      report_contents:
        items:
          $ref: '#/definitions/10_0_0_50_tuan_quang_tran_vms-v2_internal_models.ContentReport'
        type: array
      total:
        type: integer
    type: object
  internal_controllers.GetReportContentListResponse:
    properties:
      data:
        $ref: '#/definitions/internal_controllers.GetReportContentListData'
      status:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aioz Stream Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aioz Stream API
  version: "1.0"
paths:
...
  /reports:
    get:
      consumes:
        - application/json
      description: Get a list of report contents.
      parameters:
        - in: query
          name: limit
          type: integer
        - in: query
          name: offset
          type: integer
        - in: query
          name: order_by
          type: string
        - in: query
          name: sort_by
          type: string
        - in: query
          name: status
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.GetReportContentListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Get a list of report contents
      tags:
        - reports
    post:
      consumes:
        - application/json
      description: Create a report content.
      parameters:
        - description: Create report content request
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/internal_controllers.CreateReportContentRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Create a report content
      tags:
        - reports
  /reports/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a report content.
      parameters:
        - description: Report ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Delete a report content
      tags:
        - reports
    patch:
      consumes:
        - application/json
      description: Update a report content.
      parameters:
        - in: query
          name: status
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Update a report content
      tags:
        - reports
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"